Removed warnings by explicitly casting to `u8` for controllers

+++ Source/Core/Core/HW/WiimoteEmu/Attachment/Turntable.cpp
@@ -65,8 +65,8 @@ void Turntable::GetState(u8* const data)
 	double x, y;
 	m_stick->GetState(&x, &y);
 
-	ttdata->sx = (x * 0x1F) + 0x20;
-	ttdata->sy = (y * 0x1F) + 0x20;
+	ttdata->sx = (u8)(x * 0x1F) + 0x20;
+	ttdata->sy = (u8)(y * 0x1F) + 0x20;
 	}
 
 	// left table
@@ -75,7 +75,7 @@ void Turntable::GetState(u8* const data)
 	s8 tt_;
 	m_left_table->GetState(&tt);
 
-	tt_ = tt * 0x1F;
+	tt_ = (s8)(tt * 0x1F);
 
 	ttdata->ltable1 = tt_;
 	ttdata->ltable2 = tt_ >> 5;
@@ -87,7 +87,7 @@ void Turntable::GetState(u8* const data)
 	s8 tt_;
 	m_right_table->GetState(&tt);
 
-	tt_ = tt * 0x1F;
+	tt_ = (s8)(tt * 0x1F);
 
 	ttdata->rtable1 = tt_;
 	ttdata->rtable2 = tt_ >> 1;
@@ -101,7 +101,7 @@ void Turntable::GetState(u8* const data)
 	u8 dial_;
 	m_effect_dial->GetState(&dial);
 
-	dial_ = dial * 0x0F;
+	dial_ = (u8)(dial * 0x0F);
 
 	ttdata->dial1 = dial_;
 	ttdata->dial2 = dial_ >> 3;
@@ -112,7 +112,7 @@ void Turntable::GetState(u8* const data)
 	double cfs;
 	m_crossfade->GetState(&cfs);
 
-	ttdata->slider = (cfs * 0x07) + 0x08;
+	ttdata->slider = (u8)(cfs * 0x07) + 0x08;
 	}
 
 	// buttons
